requirements.txt
albumentations==1.3.1
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
cycler==0.11.0
Cython==3.0.10
dnspython==2.3.0
docutils==0.20.1
easydict==1.13
flatbuffers==24.3.25
fonttools==4.38.0
humanfriendly==10.0
idna==3.7
imageio==2.31.2
importlib-metadata==6.7.0
insightface==0.7.3
joblib==1.3.2
Kivy==2.3.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
matplotlib==3.5.3
mpmath==1.3.0
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
onnx==1.14.1
onnxruntime==1.14.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
Pillow==9.5.0
prettytable==3.7.0
protobuf==4.24.4
Pygments==2.17.2
pymongo==4.7.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
PyWavelets==1.3.0
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sympy==1.10.1
tensorboardX==2.6.2.2
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.4
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
zipp==3.15.0
